fix replace code
