+++ src/27-replace/.gitignore
@@ -6,4 +6,4 @@ package.json
 package.yaml
 ecli
 bootstrap
-textreplace2
+replace
